+++ apps/sample-react-dapp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # sample-react-dapp
 
+## 0.0.2-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @agateh/solana-headless-react@1.0.2-next.0
+
 ## 0.0.1
 
 ### Patch Changes